+++ rusqlite/src/load_extension_guard.rs
@@ -1,13 +1,13 @@
-use {SqliteResult, Connection};
+use {Result, Connection};
 
 /// RAII guard temporarily enabling SQLite extensions to be loaded.
 ///
 /// ## Example
 ///
 /// ```rust,no_run
-/// # use rusqlite::{Connection, SqliteResult, SqliteLoadExtensionGuard};
+/// # use rusqlite::{Connection, Result, SqliteLoadExtensionGuard};
 /// # use std::path::{Path};
-/// fn load_my_extension(conn: &Connection) -> SqliteResult<()> {
+/// fn load_my_extension(conn: &Connection) -> Result<()> {
 ///     let _guard = try!(SqliteLoadExtensionGuard::new(conn));
 ///
 ///     conn.load_extension(Path::new("my_sqlite_extension"), None)
@@ -20,7 +20,7 @@ pub struct SqliteLoadExtensionGuard<'conn> {
 impl<'conn> SqliteLoadExtensionGuard<'conn> {
     /// Attempt to enable loading extensions. Loading extensions will be disabled when this
     /// guard goes out of scope. Cannot be meaningfully nested.
-    pub fn new(conn: &Connection) -> SqliteResult<SqliteLoadExtensionGuard> {
+    pub fn new(conn: &Connection) -> Result<SqliteLoadExtensionGuard> {
         conn.load_extension_enable().map(|_| SqliteLoadExtensionGuard { conn: conn })
     }
 }